1.6

+++ src/Index.java
@@ -1,22 +1,22 @@
 
 public class Index {
-	private int start;
-	private int length;
+	private long start;
+	private long length;
 	private int line;
 	private int cont;
 	
-	public Index(int start, int length, int line, int cont){
+	public Index(long start, long length, int line, int cont){
 		this.start = start;
 		this.length = length;
 		this.line = line;
 		this.cont = cont;
 	}
 	
-	public int getStart(){
+	public long getStart(){
 		return start;
 	}
 	
-	public int getLength(){
+	public long getLength(){
 		return length;
 	}
